gpt_backend/admin_service/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
pytest
httpx
pyyaml
requests
structlog
python-json-logger
pyjwt
sqlalchemy
pymongo